chore(fe): add carousel story and reset index.vue
